Move DefaultFileMetadata to impl package
