+++ subprojects/build-cache-packaging/src/test/groovy/org/gradle/caching/internal/command/BuildCacheCommandFactoryTest.groovy
@@ -27,8 +27,8 @@ import org.gradle.caching.internal.origin.OriginMetadataFactory
 import org.gradle.caching.internal.origin.OriginReader
 import org.gradle.caching.internal.origin.OriginWriter
 import org.gradle.caching.internal.packaging.BuildCacheEntryPacker
-import org.gradle.internal.file.DefaultFileMetadata
 import org.gradle.internal.file.TreeType
+import org.gradle.internal.file.impl.DefaultFileMetadata
 import org.gradle.internal.hash.HashCode
 import org.gradle.internal.snapshot.DirectorySnapshot
 import org.gradle.internal.snapshot.FileMetadata
