+++ subprojects/native/src/main/java/org/gradle/internal/nativeintegration/filesystem/services/FallbackFileMetadataAccessor.java
@@ -16,8 +16,8 @@
 
 package org.gradle.internal.nativeintegration.filesystem.services;
 
-import org.gradle.internal.file.DefaultFileMetadata;
 import org.gradle.internal.file.FileMetadataSnapshot;
+import org.gradle.internal.file.impl.DefaultFileMetadata;
 import org.gradle.internal.nativeintegration.filesystem.FileMetadataAccessor;
 
 import java.io.File;
