+++ subprojects/snapshots/src/test/groovy/org/gradle/internal/snapshot/FileSystemSnapshotBuilderTest.groovy
@@ -17,8 +17,8 @@
 package org.gradle.internal.snapshot
 
 import org.gradle.api.internal.cache.StringInterner
-import org.gradle.internal.file.DefaultFileMetadata
 import org.gradle.internal.file.FileType
+import org.gradle.internal.file.impl.DefaultFileMetadata
 import org.gradle.internal.hash.FileHasher
 import org.gradle.internal.hash.HashCode
 import org.gradle.util.TextUtil
